modificiación de memset y strlen

+++ libft/ft_strlen.c
@@ -6,16 +6,16 @@
 /*   By: jaojeda- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/07 20:10:25 by jaojeda-          #+#    #+#             */
-/*   Updated: 2025/04/07 20:59:49 by jaojeda-         ###   ########.fr       */
+/*   Updated: 2025/04/08 15:03:45 by jaojeda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 /*
 #include <stdio.h>
 #include <string.h>
 */
-unsigned int	ft_strlen(const char	*str)
+size_t	ft_strlen(const char	*str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
